pyproject.toml
[tool.poetry]
name = "rag-eval"
version = "0.1.0"
description = ""
authors = ["rozepatricia <patricia.roze@hec.edu>"]
readme = "README.md"
packages = [{include = "rag_eval"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
langchain-openai = "^0.0.2.post1"
chromadb = "^0.4.22"
python-dotenv = "^1.0.0"
pre-commit = "^3.6.0"
langchain = "^0.1.1"
langchainhub = "^0.1.14"
argparse = "^1.4.0"
logging = "^0.4.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
